Refactor: change meta titles and description and add header tags for titles
